Clean engine :
* Move some logic to the engine.
* :retab python files
uses spaces over tabs (pep8)
* use yaml configuration file.
