--- reservation.json
@@ -1,20 +1,0 @@
-{
-	"resources": {
-		"parapluie": 4
-	},
-
-	"subnets": {
-		"rennes": "slash_22=1"
-	},
-
-	"openstack": {
-		"controllers": 1,
-		"compute":		1,
-		"network":		1,
-		"storage":		1
-	},
-
-	"walltime": "03:00:00",
-
-	"job_name": "discovery-kolla"
-}
